requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
woocommerce>=3.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
deep-translator>=1.11.4
